feature - added basic structure of header, navigation and footer
