+++ frontend/src/components/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer: React.FC = () => {
+  return (
+    <footer className="bg-gray-800 text-white text-center p-4 mt-auto">
+      <p>&copy; {new Date().getFullYear()} Developed by Ismael Ponce</p>
+    </footer>
+  );
+};
+
+export default Footer;
