+++ frontend/src/components/Header.tsx
@@ -1,0 +1,9 @@
+const Header: React.FC = () => {
+  return (
+    <header className="bg-blue-500 text-white text-center p-4 sticky top-0 z-50 shadow-md">
+      <h1>Movies</h1>
+    </header>
+  );
+};
+
+export default Header;
