+++ frontend/src/pages/Principals.tsx
@@ -1,0 +1,5 @@
+const Principals: React.FC = () => {
+  return <div>Principals</div>;
+};
+
+export default Principals;
